moved draggable and droppable folder inside DnDQuestion
