+++ apps/frontend/src/components/Questions/DnDQuestion/Draggable/Draggable.tsx
