+++ apps/frontend/src/components/Questions/DnDQuestion/Droppable/Droppable.tsx
