+++ apps/frontend/src/components/Questions/DnDQuestion/Droppable/index.ts
